# docs/swagger.yaml
definitions:
  application.Application:
    properties:
      app_hash:
        description: 应用hash
        type: string
      created_at:
        type: string
      desc:
        description: 应用描述
        type: string
      icon:
        description: icon
        type: string
      id:
        type: integer
      name:
        description: 应用名称
        type: string
      status:
        description: 应用状态 0-未发布 1-上线 2-下线
        type: integer
      updated_at:
        type: string
      version_id:
        description: 当前版本id
        type: integer
    type: object
...
  server.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
  server.AppArg:
    properties:
      description:
        description: 应用描述
        type: string
      icon:
        description: 应用icon
        type: string
      name:
        description: 应用名称
        type: string
    required:
    - description
    - icon
    - name
    type: object
  server.AppListVO:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/application.Application'
        type: array
    type: object
  server.AppUpdateArg:
    properties:
      description:
        description: 应用描述
        type: string
      icon:
        description: 应用icon
        type: string
      id:
        type: integer
      name:
        description: 应用名称
        type: string
    required:
    - description
    - icon
    - id
    - name
    type: object
  server.ApplicationByHashArgs:
    properties:
      hash:
        description: 应用hash
        type: string
    required:
    - hash
    type: object
...
  server.ClientUserItem:
    properties:
      id:
        description: 用户id
        type: integer
      true_name:
        description: 真实姓名
        type: string
    type: object
  server.ClientUserListVO:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/server.ClientUserItem'
        type: array
    type: object
...
  server.IdStruct:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  server.Permission:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  server.UserInfoVO:
    properties:
      app_id:
        description: 应用id
        type: integer
      app_name:
        description: 应用名称
        type: string
      id:
        description: 用户id
        type: integer
      mobile:
        description: 手机号
        type: string
      nickname:
        description: 昵称
        type: string
      organization_name:
        description: 组织名称
        type: string
      true_name:
        description: 真实姓名
        type: string
    type: object
  server.UserListArg:
    properties:
      group_id:
        description: 组织id
        type: integer
      limit:
        description: 页容量
        type: integer
      name:
        description: 用户名称
        type: string
      page:
        description: 页码
        type: integer
      roles:
        description: 角色集合
        items:
          type: integer
        type: array
    required:
    - limit
    - page
    type: object
  server.UserListItem:
    properties:
      account:
        description: 账号/用户名
        type: string
      created_at:
        description: 创建时间
        type: string
      group_name:
        description: 组织
        type: string
      id:
        type: integer
      nickname:
        description: 昵称
        type: string
      role_name:
        description: 角色
        items:
          type: string
        type: array
      true_name:
        description: 真实姓名
        type: string
      updated_at:
        description: 更新时间
        type: string
    type: object
  server.UserListVO:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/server.UserListItem'
        type: array
    type: object
  server.UserLoginArg:
    properties:
      account:
        type: string
      app_hash:
        type: string
      password:
        type: string
    required:
    - account
    - app_hash
    - password
    type: object
  server.UserPermission:
    properties:
      child:
        items:
          $ref: '#/definitions/server.Permission'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
  server.UserPwdArg:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
...
info:
  contact: {}
paths:
...
  /api/hash:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 请求体
        in: body
        name: ApplicationByHashArgs
        required: true
        schema:
          $ref: '#/definitions/server.ApplicationByHashArgs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/application.Application'
              type: object
      tags:
      - Application
...
  /api/user/info:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: Bearer +token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.UserInfoVO'
              type: object
      summary: 用户信息
      tags:
      - client/user
  /api/user/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: Bearer +token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用户名称
        in: query
        name: name
        type: string
      - description: 组织id,多个用','分隔
        in: query
        name: org_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.ClientUserListVO'
              type: object
      summary: 用户列表[update]
      tags:
      - client/user
  /api/user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用hash
        in: header
        name: app_hash
        required: true
        type: string
      - description: 请求体
        in: body
        name: UserLoginArg
        required: true
        schema:
          $ref: '#/definitions/server.UserLoginArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: string
              type: object
      summary: 用户登录
      tags:
      - client/user
  /api/user/password/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: Bearer +token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求体
        in: body
        name: UserPwdArg
        required: true
        schema:
          $ref: '#/definitions/server.UserPwdArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 用户修改密码
      tags:
      - client/user
  /api/user/permission:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: Bearer +token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  items:
                    $ref: '#/definitions/server.UserPermission'
                  type: array
              type: object
      summary: 用户权限
      tags:
      - client/user
  /app/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/application.Application'
              type: object
      summary: 应用详情
      tags:
      - Application
  /app/edit:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 请求体
        in: body
        name: AppUpdateArg
        required: true
        schema:
          $ref: '#/definitions/server.AppUpdateArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 编辑应用
      tags:
      - Application
  /app/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.AppListVO'
              type: object
      summary: 应用列表
      tags:
      - Application
  /app/new:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 请求体
        in: body
        name: AppArg
        required: true
        schema:
          $ref: '#/definitions/server.AppArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 新增应用
      tags:
      - Application
...
  /user/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 请求体
        in: body
        name: UserListArg
        required: true
        schema:
          $ref: '#/definitions/server.UserListArg'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  $ref: '#/definitions/server.UserListVO'
              type: object
      summary: 用户列表
      tags:
      - user
...
  /user/password/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: 语言(zh/en),默认en
        in: header
        name: lang
        type: string
      - description: 应用id
        in: header
        name: app-id
        required: true
        type: string
      - description: 请求体
        in: body
        name: IdStruct
        required: true
        schema:
          $ref: '#/definitions/server.IdStruct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server.ApiResponse'
            - properties:
                result:
                  type: object
              type: object
      summary: 用户密码重置
      tags:
      - user
...
swagger: "2.0"